+++ CHANGELOG.md
@@ -23,11 +23,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   `translateCtxAndChildren()`.
 - Properties `parser`, `lexer`, `errorHandler` and `tokenStream` inside the class `KipperFileContext`.
 - File `built-ins.ts`, which defines the behaviour on how to define built-in items inside a kipper program. This
-  primarily includes global functions, which can be represented using the interface `GlobalFunction`. (In work!)
+  primarily includes global functions, which can be represented using the interface `Function`. (In work!)
 - Implemented `**` (Power-to) as a valid arithmetic expression.
 - Implemented `RuntimeCompileConfig` and `CompileConfig`, which may be passed onto `KipperCompile.compile()` to configure
   the compilation behaviour.
 - Implemented new module `/compiler/tokens`, which contains the parse token implementations.
+- Implemented basic global function `print` that will be available inside a Kipper program per default (unless 
+  forcibly changed).
 
 ### Changed
 - Made return of `KipperCompiler.parse()` to `KipperFileContext`, which wraps the generated parse tree.